# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRMS REST API",
        "description": "Human Resource Management System API Documentation",
        "contact": {
            "email": "admin@hrms.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/jobs": {
            "get": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Get list of job postings",
                "operationId": "7496e064a1963b40b600bfc0432fdd1d",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (open/closed)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Create job posting",
                "operationId": "ba4f0543150edc94b540a07dadb111b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "position",
                                    "description"
                                ],
                                "properties": {
                                    "position": {
                                        "type": "string",
                                        "example": "Senior Software Engineer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "requirements": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Mumbai"
                                    },
                                    "salary_range": {
                                        "type": "string",
                                        "example": "10-15 LPA"
                                    },
                                    "employment_type": {
                                        "type": "string",
                                        "example": "Full-time"
                                    },
                                    "closing_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/jobs/{id}/applications": {
            "get": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Get applications for a job",
                "operationId": "d8e07686f4edc32b5f759c20c2a756dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/jobs/{id}/apply": {
            "post": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Apply for a job",
                "operationId": "473ab3d2e13f8b134c0057813f5f4e56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "applicant_name",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "applicant_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "resume": {
                                        "type": "string"
                                    },
                                    "cover_letter": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Application submitted"
                    }
                }
            }
        },
        "/api/v1/applications/{id}/status": {
            "put": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Update application status",
                "operationId": "10af97a063de8adf08258cfdaa3b7ce7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "shortlisted"
                                    },
                                    "interview_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "remarks": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}